GuidedProposals.jl

test documentation

GuidedProposals.AbstractGuidingTermSolver

GuidedProposals.GuidProp

GuidedProposals.HFcBuffer

GuidedProposals.HFcContainer

GuidedProposals.HFcSolver

Base.Libc.time

GuidedProposals.F

GuidedProposals.F

GuidedProposals.H

GuidedProposals.H

GuidedProposals.HFc0

GuidedProposals.HFc0

GuidedProposals.c

GuidedProposals.c

GuidedProposals.fetch_eltype

GuidedProposals.fetch_xT_plus

GuidedProposals.get_or_default

GuidedProposals.init_solver

GuidedProposals.outer

GuidedProposals.prepare_static_saving_types

GuidedProposals.reformat

GuidedProposals.size_of_HFc_buffer

GuidedProposals.size_of_HFc_solution

GuidedProposals.static_accessor_HFc

GuidedProposals.update_HFc

GuidedProposals.update_HFc!

GuidedProposals.GuidProp — Type

struct GuidProp{
        K,DP,DW,SS,R,R2,O,S,T
        } <: DiffusionDefinition.DiffusionProcess{K,DP,DW,SS}
    P_target::R
    P_aux::R2
    obs::O
    guiding_term_solver::S
end

Struct defining guided proposals of M Schauer, F van der Meulen and H van Zanten. See Mider M, Schauer M and van der Meulen F Continuous-discrete smoothing of diffusions (2020) for a comprehensive overview of the mathematics behind this object. It computes and stores the guiding term ∇logρ and allows for simulation of guided proposals and computation of their likelihood.

    GuidProp(
            tt,
            P_target::R,
            P_aux::R2,
            obs::O,
            solver_choice=(
                solver=Tsit5(),
                ode_type=:HFc,
                convert_to_HFc=false,
                mode=:inplace,
                gradients=false,
                eltype=Float64,
            ),
            next_guiding_term=nothing
        ) where {R,R2,O}

Default constructor. `P_target` and `P_aux` are the target and auxiliary
diffusion laws respectively. `tt` is the time-grid on which `∇logρ` needs
to be computed. `obs` is the terminal observation (and the only one on the
interval (`tt[1]`, `tt[end]`]). `solver_choice` specifies the type of ODE
solver that is to be used for computations of `∇logρ`
    ( it is a `NamedTuple`, where `solver` specifies the algorithm for
    solving ODEs (see the documentation of DifferentialEquations.jl for
    possible choices), `ode_type` picks the ODE system (between :HFc, :MLμ
    and :Pν), `convert_to_HFc` indicates whether to translate the results of
    M,L,μ solver to H,F,c objects, `mode` is a flag indicating the way in
    which data is being handled:
    - `:inplace`: uses regular arrays to store the data (requires functions
                  B!, β!, σ! and a! to be defined)
    - `:outofplace`: operates on static arrays
    - `:gpu`: operates on GPU arrays [TODO not implemented yet]
    `gradients` is a flag indicating whether automatic differentiation is to
    be employed and `eltype` indicates the data-type of each container's
    member. )
Finally, `next_guiding_term` is the guided proposal for the subsequent
inter-observation interval.

GuidedProposals.F — Method

F(P::GuidProp, i)

Return saved vector Fi.

GuidedProposals.F — Method

F(s::HFcSolver, i::Integer)

Return saved vector Fi.

GuidedProposals.H — Method

H(P::GuidProp, i)

Return saved matrix Hi.

GuidedProposals.H — Method

H(s::HFcSolver, i::Integer)

Return saved matrix Hi.

GuidedProposals.c — Method

c(P::GuidProp, i)

Return saved scalar ci.

GuidedProposals.c — Method

c(s::HFcSolver, i::Integer)

Return saved scalar ci.

GuidedProposals.AbstractGuidingTermSolver — Type

AbstractGuidingTermSolver{Tmode}

Supertype for ODE solvers (solving H,F,c system or M,L,μ system or P,ν system). Tmode is a flag for whether computations are done in-place (with states represented by vectors), out-of-place (with state represented by StaticArrays), or on GPUs (with states represented by cuArrays).

GuidedProposals.HFcBuffer — Type

struct HFcBuffer{
        T,D,TB,Tβ,Tσ,Ta,Tmat,Tvec
        } <: DiffusionDefinition.AbstractBuffer{T}
    data::Vector{T}
    B::TB
    β::Tβ
    σ::Tσ
    a::Ta
    mat::Tmat
    vec::Tvec
end

A buffer for temporary computations of in-place ODE solvers solving for H,F,c system.

GuidedProposals.HFcContainer — Type

struct HFcContainer{T,D,TH,TF,Tc} <: AbstractBuffer{T}
    data::Vector{T}
    H::TH
    F::TF
    c::Tc
end

A buffer containing data for in-place computations of H,F,c terms.

GuidedProposals.HFcSolver — Type

struct HFcSolver{Tmode,Tsv,Tps,Tcb,Ts,T,Ta} <: AbstractGuidingTermSolver{Tmode}
    saved_values::Tsv
    problem_setup::Tps
    callback::Tcb
    solver::Ts
    HFc0::T
    access::Ta
end

Struct for solving a (H,F,c)-system of ODEs.

    HFcSolver(
        ::Val{:inplace},
        tt,
        xT_plus,
        P,
        obs,
        choices,
    )
Constructor for an ODE solver with in-place operations. Pre-allocates space
and solves a (H,F,c)-system of ODEs once, backward in time on the interval
`(tt[1], tt[end])`, with a terminal condition computed from `xT_plus`. `P`
is the auxiliary diffusion law, `obs` is the observation made at time
`tt[end]` and `choices` contains information about eltype and the algorithm
for solving ODEs. H,F,c are saved on a grid of time-points `tt`.

    HFcSolver(
        ::Val{:outofplace},
        tt,
        xT_plus,
        P,
        obs,
        choices
    )
Constructor for an ODE solver with out-of-place operations using
StaticArrays. Initialises the object and solves a (H,F,c)-system of ODEs
once, backward in time on the interval `(tt[1], tt[end])`, with a terminal
condition computed from `xT_plus`. `P` is the auxiliary diffusion law, `obs`
is the observation made at time `tt[end]` and `choices` contains information
about eltype and the algorithm for solving ODEs. H,F,c are saved on a grid
of time-points `tt`.

Base.Libc.time — Method

Base.time(P::GuidProp, i)

Return time-point tt[i] corresponding to a saved state of ODEs (with tt[1] indicating time 0+ and tt[end] indicating time T).

GuidedProposals.HFc0 — Method

HFc0(P::GuidProp)

Return the container with data that can be used to reconstruct H,F,c evaluated at time 0+ for the guided proposal P.

GuidedProposals.HFc0 — Method

HFc0(s::HFcSolver)

Return the data containing H,F,c terms for the time 0+

GuidedProposals.fetch_eltype — Method

fetch_eltype(choices, P)

Determine the type of the elements that is supposed to be used by the internal containers of this package. If choice.gradients flag is turned on, then use the same type as the eltypes of the parameters in the auxiliary law. Otherwise, use the type specified in choice. If neither the choice.gradients flag is on, nor a default is provided, use Float64.

GuidedProposals.fetch_xT_plus — Function

fetch_xT_plus(::Val{:inplace}, next_guiding_term, el, dim_of_proc)

If this is not the last inter-observation interval, fetch the data containing H,F,c computed for the left time-limit of the subsequent interval. Otherwise, instantiate a zero-term.

GuidedProposals.get_or_default — Method

get_or_default(container, elem::Symbol, default)

Return container.elem if it exists, otherwise return default

GuidedProposals.init_solver — Method

init_solver(
    ::Val{:hfc},
    ::Any,
    tt,
    P_aux,
    obs,
    mode::Val,
    choices,
    next_guiding_term
)

Initialise ODE solver for H,F,c, preallocate space and solve it once. tt is the time-grid on which ∇logρ is to be saved. P_aux is the auxiliary law, obs is the terminal observation, mode is for differentiating between in-place, out-of-place and gpu constructors for the guiding term solver, choices contains additional information that is passed on (and which is about eltype and a chosen algorithm for the ODE solvers) and finally, next_guiding_term is the guided proposal used on the subsequent inter-observation interval.

GuidedProposals.outer — Method

outer(x)

Compute an outer product

GuidedProposals.prepare_static_saving_types — Method

prepare_static_saving_types(::Val{:hfc}, ::Val{D}, el) where D

Define data-types for H,F,c computed by out-of-place solver that are to be saved internally.

GuidedProposals.reformat — Method

reformat(solver_choice::NamedTuple, last_interval::Bool, P_aux)

Re-format the solver_choice by splitting it into two NamedTuples and populating any missing entries with defaults. P_aux is the law of the auxiliary diffusion that is needed in case a flag for computing gradients is turned on.

GuidedProposals.size_of_HFc_buffer — Method

size_of_HFc_solution(d)

Length of a vector containing temporary data needed for in-place solver of H,F,c, when the underlying process has dimension d.

GuidedProposals.size_of_HFc_solution — Method

size_of_HFc_solution(d)

Compute the size of a vector containing H,F,c elements

GuidedProposals.static_accessor_HFc — Method

static_accessor_HFc(u::SVector, ::Val{T}) where T

Access data stored in the container u so that it matches the shapes of H,F,c and points to the correct points in u. T is the dimension of the stochastic process.

GuidedProposals.update_HFc! — Method

update_HFc!(u_T, u_Tplus, obs, access)

Update equations for H,F,c at the times of observations. Save the data into u_T.

GuidedProposals.update_HFc — Method

update_HFc(u_Tplus, obs, access)

Update equations for H,F,c at the times of observations.