GuidedProposals.jl

test documentation

GuidedProposals.GuidProp

GuidedProposals.HFcSolver

GuidedProposals.HFc0

GuidedProposals._B

GuidedProposals._F

GuidedProposals._H

GuidedProposals._L

GuidedProposals._M

GuidedProposals._P

GuidedProposals._a

GuidedProposals._c

GuidedProposals._temp_matH

GuidedProposals._temp_vecH

GuidedProposals._β

GuidedProposals._μ

GuidedProposals._ν

GuidedProposals._σ

GuidedProposals.init_hfc

GuidedProposals.init_hfc

GuidedProposals.init_solvers

GuidedProposals.outer

GuidedProposals.update_HFc!

GuidedProposals.GuidProp — Type

GuidProp{R,R2,O,C,SM,SH,SP}

Struct defining guided proposals of M Schauer, F van der Meulen and H van Zanten. See Mider M, Schauer M and van der Meulen F Continuous-discrete smoothing of diffusions (2020) for a comprehensive overview of the mathematics behind this object. It computes and stores the guiding term ∇logρ and allows for simulation of guided proposals and computation of their likelihood.

    GuidProp(
        tt,
        P_target::R,
        P_aux::R2,
        obs::O,
        solver_choice=(
            solver=:Tsit5,
            ode_type=:HFc,
            convert_to_HFc=false,
            inplace=false,
            save_as_type=nothing,
            ode_data_type=nothing,
        ),
        next_guiding_term=nothing
    ) where {R,R2,O}

Default constructor. `P_target` and `P_aux` are the target and auxiliary
diffusion laws respectively. `tt` is the time-grid on which `∇logρ` needs
to be computed. `obs` is the terminal observation (and the only one on the
interval (`tt[1]`, `tt[end]`]). `solver_choice` specifies the type of ODE
solver that is to be used for computations of `∇logρ` (it is a `NamedTuple`,
with `solver` specifying the ODE solver, `ode_type` picking the ODE system
(between :HFc, :MLμ and :Pν), `convert_to_HFc` indicating whether to
translate the results of M,L,μ solver to H,F,c objects, `inplace` is a flag
for whether to use in-place ODE solvers (for which B!, β!, σ! and a!
functions need to be defined for the auxiliary diffusion), `save_as_type`
specifies the datatypes in which the terms computed for `∇logρ` are to be
saved and `ode_data_type` specifies the data type used for internal
computations of ODE solvers). Finally, `next_guiding_term` is the guided
proposal for the subsequent inter-observation interval.

GuidedProposals.HFcSolver — Type

struct HFcSolver{Tp,Tcb,T}

Struct for solving a (H,F,c)-system of ODEs.

    HFcSolver(
        ::Val{:inplace},
        tt,
        xT_plus,
        P,
        solver_type,
        save_as_type,
    )
Constructor for an ODE solver with in-place operations. Pre-allocates space
and solves once a (H,F,c)-system of ODEs backward in time on the interval
`(tt[1], tt[end])`, with a terminal condition computed from `xT_plus`. `P`
is the auxiliary diffusion, and `solver_type` indicates the algorithm for
solving ODEs. `save_as_type` gives datatypes that H,F,c are to be saved as
and `tt` gives a grid of time-points at which they need to be saved.

    HFcSolver(
        ::Val{:outofplace},
        tt,
        xT_plus,
        P,
        solver_type,
        save_as_type,
    )
Constructor for an ODE solver with out-of-place operations.
NOTE: currently not implemented

GuidedProposals.HFc0 — Method

HFc0(s::HFcSolver)

Return the data containing H,F,c terms for the time 0+

GuidedProposals._B — Method

_B(buffer::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix B that are stored in a container buffer

GuidedProposals._F — Method

_F(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a vector F that are stored in a container data

GuidedProposals._H — Method

_H(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix H that are stored in a container data

GuidedProposals._L — Method

_L(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix L that are stored in a container data

GuidedProposals._M — Method

_M(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix M that are stored in a container data

GuidedProposals._P — Method

_P(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix P that are stored in a container data

GuidedProposals._a — Method

_a(buffer::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix a that are stored in a container buffer

GuidedProposals._c — Method

_c(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a scalar c that is stored in a container data

GuidedProposals._temp_matH — Method

_temp_matH(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a temporary matrix, stored in a container buffer. NOTE: _temp_matH and _temp_vecH access overlapping chunks of memory.

GuidedProposals._temp_vecH — Method

_temp_vecH(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a temporary vector, stored in a container buffer. NOTE: _temp_matH and _temp_vecH access overlapping chunks of memory.

GuidedProposals._β — Method

_β(buffer::AbstractArray, ::Val{T}) where T

Provide a view to contents of a vector β that are stored in a container buffer

GuidedProposals._μ — Method

_μ(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a vector μ that are stored in a container data

GuidedProposals._ν — Method

_ν(data::AbstractArray, ::Val{T}) where T

Provide a view to contents of a vector ν that are stored in a container data

GuidedProposals._σ — Method

_σ(buffer::AbstractArray, ::Val{T}) where T

Provide a view to contents of a matrix σ that are stored in a container buffer

GuidedProposals.init_hfc — Method

init_hfc(::Val{false}, ...)

Nothing to initialise if a flag for HFc solver is turned off

GuidedProposals.init_hfc — Method

init_hfc(::Val{true}, tt, P_aux, obs, choice, next_guiding_term)

Initialise ODE solver for H,F,c, preallocate space and solve it once. tt is the time-grid on which ∇logρ is to be saved. P_aux is the auxiliary law, obs is the terminal observation, choice is the NamedTuple with various choices for the the ODE solver (here, the ones used are: solver–-the type of ODE solver, inplace–-whether to perform computations in-place, save_as_type–-the datatypes in which H,F,c are to be solved (if set to nothing, then Matrix{Float64}, Vector{Float64} and Float64 are used by default for storing H, F and c respectively), ode_data_type–-the datatypes in which the ODE solvers should store all the data (if set to nothing, then Vector{Float64} is used by default)). Finally, next_guiding_term is the guided proposal used on the subsequent inter-observation interval.

GuidedProposals.init_solvers — Method

init_solvers(params)

Initialise appropriate ODE solvers, according to how they are specified in params. Validity of the format of params is not checked, because it is only ever called by hard-coded constructor of GuidProp.

GuidedProposals.outer — Method

outer(x)

Compute an outer product

GuidedProposals.update_HFc! — Method

update_HFc!(u_T, u_Tplus, obs, access)

Update equations for H,F,c at the times of observations.