Index

GuidedProposals.AbstractGuidingTermSolver

GuidedProposals.GuidProp

GuidedProposals.GuidProp

GuidedProposals.HFcBuffer

GuidedProposals.HFcContainer

GuidedProposals.HFcSolver

GuidedProposals.IntegrationRule

GuidedProposals.LeftRule

Base.Libc.time

DiffusionDefinition.constdiff

DiffusionDefinition.dimension

GuidedProposals.F

GuidedProposals.F

GuidedProposals.H

GuidedProposals.H

GuidedProposals.HFc0

GuidedProposals.HFc0

GuidedProposals.c

GuidedProposals.c

GuidedProposals.fetch_eltype

GuidedProposals.fetch_xT_plus

GuidedProposals.get_or_default

GuidedProposals.init_solver

GuidedProposals.mode

GuidedProposals.outer

GuidedProposals.prepare_static_saving_types

GuidedProposals.recompute_guiding_term!

GuidedProposals.reformat

GuidedProposals.size_of_HFc_buffer

GuidedProposals.size_of_HFc_solution

GuidedProposals.static_accessor_HFc

GuidedProposals.update_HFc

GuidedProposals.update_HFc!

GuidedProposals.GuidProp — Method

GuidProp(P::GuidProp, θ°, η°)

Create new guided proposals object with new parameters θ° parametrizing the diffusion laws and η° parametrizing the observations. Keep the pre-allocated spaces for solvers unchanged.

GuidedProposals.F — Method

F(P::GuidProp, i)

Return saved vector Fi.

GuidedProposals.F — Method

F(s::HFcSolver, i::Integer)

Return saved vector Fi.

GuidedProposals.H — Method

H(P::GuidProp, i)

Return saved matrix Hi.

GuidedProposals.H — Method

H(s::HFcSolver, i::Integer)

Return saved matrix Hi.

GuidedProposals.c — Method

c(P::GuidProp, i)

Return saved scalar ci.

GuidedProposals.c — Method

c(s::HFcSolver, i::Integer)

Return saved scalar ci.

GuidedProposals.recompute_guiding_term! — Function

recompute_guiding_term!(P::GuidProp, next_guided_prop=nothing)

Recompute the guiding term (most often used after update of parameters or change of an observation). next_guided_prop is the guided proposal law from the subsequent interval

GuidedProposals.AbstractGuidingTermSolver — Type

AbstractGuidingTermSolver{Tmode}

Supertype for ODE solvers (solving H,F,c system or M,L,μ system or P,ν system). Tmode is a flag for whether computations are done in-place (with states represented by vectors), out-of-place (with state represented by StaticArrays), or on GPUs (with states represented by cuArrays).

GuidedProposals.HFcBuffer — Type

struct HFcBuffer{
        T,D,TB,Tβ,Tσ,Ta,Tmat,Tvec
        } <: DiffusionDefinition.AbstractBuffer{T}
    data::Vector{T}
    B::TB
    β::Tβ
    σ::Tσ
    a::Ta
    mat::Tmat
    vec::Tvec
end

A buffer for temporary computations of in-place ODE solvers solving for H,F,c system.

GuidedProposals.HFcContainer — Type

struct HFcContainer{T,D,TH,TF,Tc} <: AbstractBuffer{T}
    data::Vector{T}
    H::TH
    F::TF
    c::Tc
end

A buffer containing data for in-place computations of H,F,c terms.

GuidedProposals.HFcSolver — Type

struct HFcSolver{Tmode,Tsv,Tps,Tcb,Ts,T,Ta} <: AbstractGuidingTermSolver{Tmode}
    saved_values::Tsv
    problem_setup::Tps
    callback::Tcb
    solver::Ts
    HFc0::T
    access::Ta
end

Struct for solving a (H,F,c)-system of ODEs.

    HFcSolver(
        ::Val{:inplace},
        tt,
        xT_plus,
        P,
        obs,
        choices,
    )
Constructor for an ODE solver with in-place operations. Pre-allocates space
and solves a (H,F,c)-system of ODEs once, backward in time on the interval
`(tt[1], tt[end])`, with a terminal condition computed from `xT_plus`. `P`
is the auxiliary diffusion law, `obs` is the observation made at time
`tt[end]` and `choices` contains information about eltype and the algorithm
for solving ODEs. H,F,c are saved on a grid of time-points `tt`.

    HFcSolver(
        ::Val{:outofplace},
        tt,
        xT_plus,
        P,
        obs,
        choices
    )
Constructor for an ODE solver with out-of-place operations using
StaticArrays. Initialises the object and solves a (H,F,c)-system of ODEs
once, backward in time on the interval `(tt[1], tt[end])`, with a terminal
condition computed from `xT_plus`. `P` is the auxiliary diffusion law, `obs`
is the observation made at time `tt[end]` and `choices` contains information
about eltype and the algorithm for solving ODEs. H,F,c are saved on a grid
of time-points `tt`.

GuidedProposals.IntegrationRule — Type

IntegrationRule

Supertype of all integration rules. In this package we currently approximate the integrals by step functions with evaluations at the left side of the intervals.

GuidedProposals.LeftRule — Type

LeftRule <: IntegrationRule

Integration rule flag, indicating to approximate functions with step functions with height equal to function evaluation at the left side of the intevals.

Base.Libc.time — Method

Base.time(P::GuidProp, i)

Return time-point tt[i] corresponding to a saved state of ODEs (with tt[1] indicating time 0+ and tt[end] indicating time T).

DiffusionDefinition.constdiff — Method

DiffusionDefinition.dimension — Method

DD.dimension(P::GuidProp)

Dimension of the stochastic process and the driving Brownian motion (by default the same as that of the target process)

GuidedProposals.HFc0 — Method

HFc0(P::GuidProp)

Return the container with data that can be used to reconstruct H,F,c evaluated at time 0+ for the guided proposal P.

GuidedProposals.HFc0 — Method

HFc0(s::HFcSolver)

Return the data containing H,F,c terms for the time 0+

GuidedProposals.fetch_eltype — Method

fetch_eltype(choices, P)

Determine the type of the elements that is supposed to be used by the internal containers of this package. If choice.gradients flag is turned on, then use the same type as the eltypes of the parameters in the auxiliary law. Otherwise, use the type specified in choice. If neither the choice.gradients flag is on, nor a default is provided, use Float64.

GuidedProposals.fetch_xT_plus — Function

fetch_xT_plus(::Val{:inplace}, next_guided_prop, el, dim_of_proc)

If this is not the last inter-observation interval, fetch the data containing H,F,c computed for the left time-limit of the subsequent interval. Otherwise, instantiate a zero-term.

GuidedProposals.get_or_default — Method

get_or_default(container, elem::Symbol, default)

Return container.elem if it exists, otherwise return default

GuidedProposals.init_solver — Method

init_solver(
    ::Val{:hfc},
    ::Any,
    tt,
    P_aux,
    obs,
    mode::Val,
    choices,
    next_guided_prop
)

Initialise ODE solver for H,F,c, preallocate space and solve it once. tt is the time-grid on which ∇logρ is to be saved. P_aux is the auxiliary law, obs is the terminal observation, mode is for differentiating between in-place, out-of-place and gpu constructors for the guiding term solver, choices contains additional information that is passed on (and which is about eltype and a chosen algorithm for the ODE solvers) and finally, next_guided_prop is the guided proposal used on the subsequent inter-observation interval.

GuidedProposals.mode — Method

mode(P::GuidProp)

Return the mode of solving ODE systems (:inplace, :outofplace or :gpu) [TODO not used much, for multiple dispatch needs to return Val{mode}() instead, change or remove].

GuidedProposals.outer — Method

outer(x)

Compute an outer product

GuidedProposals.prepare_static_saving_types — Method

prepare_static_saving_types(::Val{:hfc}, ::Val{D}, el) where D

Define data-types for H,F,c computed by out-of-place solver that are to be saved internally.

GuidedProposals.reformat — Method

reformat(solver_choice::NamedTuple, last_interval::Bool, P_aux)

Re-format the solver_choice by splitting it into two NamedTuples and populating any missing entries with defaults. P_aux is the law of the auxiliary diffusion that is needed in case a flag for computing gradients is turned on.

GuidedProposals.size_of_HFc_buffer — Method

size_of_HFc_solution(d)

Length of a vector containing temporary data needed for in-place solver of H,F,c, when the underlying process has dimension d.

GuidedProposals.size_of_HFc_solution — Method

size_of_HFc_solution(d)

Compute the size of a vector containing H,F,c elements

GuidedProposals.static_accessor_HFc — Method

static_accessor_HFc(u::SVector, ::Val{T}) where T

Access data stored in the container u so that it matches the shapes of H,F,c and points to the correct points in u. T is the dimension of the stochastic process.

GuidedProposals.update_HFc! — Method

update_HFc!(u_T, u_Tplus, obs, access)

Update equations for H,F,c at the times of observations. Save the data into u_T.

GuidedProposals.update_HFc — Method

update_HFc(u_Tplus, obs, access)

Update equations for H,F,c at the times of observations.